New module commonh5 with fabioh5 classes that can be used in spech5
